port: LLP64 sweep -- spell every 64-bit integer via <cstdint>, never `long`
On Windows `long` is 32 bits. lang's Word::i, the engine's i64/u64, the
ext_vector_type lane types in shared/tzpl_simd.hpp, the voicer header's
standalone alias, and the bridge/app node/control/tap IDs were all `long`
and would silently truncate. All are now std::int64_t / std::uint64_t (the
same type in every namespace, which also removes the old long-vs-long-long
split between engine and lang). isize is std::ptrdiff_t (no ssize_t in the
MSVC CRT). The non-Apple SIMD branch gains the u32xN types the Apple branch
already had.

+++ synthdef-compiler/src/synthdef_compile.cpp
@@ -183,8 +183,8 @@ void runInternalAudioEngine(string dir, string synthName, int seconds) {
             // non-finite or wildly large output means the synth (or this
             // harness's fixed 2-float outlet buffer) is producing garbage.
             const char* warn = (nonFinite || maxAbs > 10.0) ? "  <<< WARNING: garbage output" : "";
-            printf("offline render: %ld frames, maxAbs=%g, finite=%s%s\n",
-                   nframes, maxAbs, nonFinite ? "NO" : "yes", warn);
+            printf("offline render: %lld frames, maxAbs=%g, finite=%s%s\n",
+                   (long long)nframes, maxAbs, nonFinite ? "NO" : "yes", warn);
         }
         free(data->outlets);
         def.funs.free(data);
